+++ api/src/everything/errors.py
