+++ api/src/everything/thumbnail.py
@@ -1,9 +1,9 @@
 import logging
 import typing
 import os
-from src.db import Media
-from src.scan import is_image
-from src.paths import get_thumbnails_directory
+from everything.db import Media
+from everything.scan import is_image
+from everything.paths import get_thumbnails_directory
 import subprocess
 
 THUMBNAIL_WIDTH = 96
